refactor(calculator): remove CalculatorRunner and give its role to Calculator
