+++ src/main/java/CalculatorApplication.java
@@ -1,13 +1,10 @@
-import stringcalculator.CalculatorRunner;
-import stringcalculator.InputView;
-import stringcalculator.OutputView;
-import stringcalculator.StringParser;
+import stringcalculator.*;
 
 public class CalculatorApplication {
     public static void main(String[] args) {
         StringParser stringParser = new StringParser();
-        CalculatorRunner calculatorRunner = new CalculatorRunner(stringParser);
-        Integer result = calculatorRunner.calculate(InputView.receiveInput());
+        Calculator calculator = new Calculator(stringParser);
+        Integer result = calculator.addString(InputView.receiveInput());
         OutputView.printResult(result);
     }
 }
