+++ src/main/java/stringcalculator/Calculator.java
@@ -3,7 +3,17 @@ package stringcalculator;
 import java.util.Arrays;
 
 public class Calculator {
-    static Integer add(Integer[] inputs) {
+    private final Parser<String> stringParser;
+
+    public Calculator(Parser<String> stringParser) {
+        this.stringParser = stringParser;
+    }
+
+    public Integer addString(String input) {
+        return this.addInteger(stringParser.parse(input));
+    }
+
+    public Integer addInteger(Integer[] inputs) {
         return Arrays.stream(inputs).reduce(0, Integer::sum);
     }
 }
