--- src/main/java/stringcalculator/CalculatorRunner.java
@@ -1,13 +1,0 @@
-package stringcalculator;
-
-public class CalculatorRunner {
-    private StringParser parser;
-
-    public CalculatorRunner(StringParser parser) {
-        this.parser = parser;
-    }
-
-    public Integer calculate(String input) {
-        return Calculator.add(parser.parse(input));
-    }
-}
